web_app/requirements_web.txt
# ============================================================
#  Virtual Staging MVP - 웹 앱 의존성
#  환경: Python 3.10+, 로컬 GPU 불필요 (Replicate 클라우드 GPU 사용)
#
#  설치 방법:
#    pip install -r requirements_web.txt
# ============================================================

# ── Streamlit 웹 프레임워크 ──────────────────────────────────
streamlit>=1.32.0

# ── 마스킹 캔버스 컴포넌트 ──────────────────────────────────
# 사용자가 브러쉬로 마스크를 그리는 핵심 UI 컴포넌트
streamlit-drawable-canvas>=0.9.3

# ── Replicate API 클라이언트 ─────────────────────────────────
# 클라우드 GPU 기반 SDXL Inpainting 실행
replicate>=0.25.0

# ── 이미지 처리 ─────────────────────────────────────────────
Pillow>=10.2.0
numpy>=1.26.0

# ── 다국어(한국어) 번역 라이브러리 ──────────────────────────
# deep-translator: Google Translate API 래퍼 (무료, API 키 불필요)
# 사용자가 입력한 한국어 프롬프트를 영어로 자동 변환
deep-translator>=1.11.4
